+++ ApiTemplate.Domain/Entities/WeatherForecast.cs
@@ -7,7 +7,7 @@ public class WeatherForecast
     public Guid Id { get; set; }
     public DateTime Date { get; set; }
 
-    public int TemperatureC { get; set; }
+    public double TemperatureC { get; set; }
 
     public string Summary { get; set; }
 }
