+++ ApiTemplate.Shared/Extensions/TemperatureExtension.cs
@@ -2,5 +2,5 @@
 
 public static class TemperatureExtension
 {
-    public static int ToFahrenheit(this int celsius) => celsius * 9 / 5 + 32;
+    public static double ToFahrenheit(this double celsius) => celsius * 9 / 5 + 32;
 }
